:necktie: Move to using a ref callback instead of an effect
